setting up serverside tests
