+++ server/jest.config.js
@@ -1,0 +1,4 @@
+export default {
+  testEnvironment: 'jest-environment-node',
+  transform: {},
+};
