+++ server/package.json
@@ -17,11 +17,14 @@
     "pg": "^8.6.0"
   },
   "devDependencies": {
-    "nodemon": "^2.0.7"
+    "@types/jest": "^27.0.1",
+    "jest": "^27.1.0",
+    "nodemon": "^2.0.7",
+    "supertest": "^6.1.6"
   },
   "scripts": {
     "start": "node src/index.js",
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js",
     "seed:admin": "node src/tasks/createAdminUser.js",
     "as:production": "set NODE_ENV=production && node src/index.js "
   },
